+++ src/main/java/com/temperatureconverter/service/TemperatureConverterService.java
@@ -18,10 +18,10 @@ public class TemperatureConverterService {
 		return repository.save(temperature);
 	}
 
-	public float celsiusToFahrenheit(String celsius) throws NumberFormatException {
+	public double celsiusToFahrenheit(String celsius) throws NumberFormatException {
 		try {
-			float c = Float.parseFloat(celsius);
-			float fahrenheit = (c * 9 / 5) + 32;
+			double c = Float.parseFloat(celsius);
+			double fahrenheit = (c * 9 / 5) + 32;
 			Temperature temperature = new Temperature(c, fahrenheit, TemperatureType.CELSIUSTOFAHRENHEIT);
 			save(temperature);
 			return fahrenheit;
@@ -29,10 +29,10 @@ public class TemperatureConverterService {
 			throw new NumberFormatException("Number Format Excepetion for input string: " + celsius );
 		}
 	}
-	public float fahrenheitToCelsius(String fahrenheit) throws NumberFormatException {
+	public double fahrenheitToCelsius(String fahrenheit) throws NumberFormatException {
 		try {
-			float f = Float.parseFloat(fahrenheit);
-			float celsius = (f - 32) * 5 / 9;
+			double f = Float.parseFloat(fahrenheit);
+			double celsius = (f - 32) * 5 / 9;
 			Temperature temperature = new Temperature(celsius, f, TemperatureType.FAHRENHEITTOCELSIUS);
 			save(temperature);
 			return celsius;
